worker/requirements.txt
faster-whisper>=1.0.0
python-Levenshtein>=0.21.0
requests>=2.31.0
